+++ GaudiKernel/tests/src/profile_Property.cpp
@@ -1,5 +1,5 @@
 /***********************************************************************************\
-* (c) Copyright 1998-2019 CERN for the benefit of the LHCb and ATLAS collaborations *
+* (c) Copyright 1998-2021 CERN for the benefit of the LHCb and ATLAS collaborations *
 *                                                                                   *
 * This software is distributed under the terms of the Apache version 2 licence,     *
 * copied verbatim in the file "LICENSE".                                            *
@@ -44,7 +44,7 @@ int main() {
 
     start = std::chrono::high_resolution_clock::now();
 
-    for ( size_t i = 0; i != N; ++i ) {
+    for ( int i = 0; i != N; ++i ) {
       props.emplace_back( names[i % NPropNames], i,
                           "some long documentation string describing what the property is for" );
     }
