+++ GaudiMP/src/Lib/TESSerializer.cpp
@@ -1,5 +1,5 @@
 /***********************************************************************************\
-* (c) Copyright 1998-2019 CERN for the benefit of the LHCb and ATLAS collaborations *
+* (c) Copyright 1998-2021 CERN for the benefit of the LHCb and ATLAS collaborations *
 *                                                                                   *
 * This software is distributed under the terms of the Apache version 2 licence,     *
 * copied verbatim in the file "LICENSE".                                            *
@@ -134,7 +134,7 @@ void GaudiMP::TESSerializer::dumpBuffer( TBufferFile& buffer ) {
 
     /* take care of links */
     LinkManager* linkMgr  = pObj->linkMgr();
-    int          numLinks = linkMgr->size();
+    long int     numLinks = linkMgr->size();
     buffer.WriteInt( numLinks );
     // now write each link
     for ( int it = 0; it != numLinks; it++ ) {
